Add files via upload
The main change i have made is to write the FitSinglets file which can be run as an executable. This seems to work for singlets but for other species I do not know how to set the initial values reliably.


To simplify the executable, I made a change to the MainLogic file to avoid having to import nmrglue. This should enable us to compile FitSinglets to a much smaller file, though I am still working on that. I think I did something with proc_bl for similar reasons.

I have temporarily broken the chemTree file slightly by disabling the chemDB section. This is because of a problem during compilation where it won't source the chemdb even when it is in the directory with the main file. Don't accept this change, but I am including it so you can compile it if you want to test it.